Demonstrate wire with nested packages
