+++ wire-di/message/message.go
@@ -1,0 +1,7 @@
+package message
+
+type Message string
+
+func NewMessage(phrase string) Message {
+	return Message(phrase)
+}
