+++ wire-di/wire.go
@@ -3,9 +3,14 @@
 
 package main
 
-import "github.com/google/wire"
+import (
+	"github.com/danhawkins/go-examples/wire/event"
+	"github.com/danhawkins/go-examples/wire/greeter"
+	"github.com/danhawkins/go-examples/wire/message"
+	"github.com/google/wire"
+)
 
-func InitializeEvent(phrase string) (Event, error) {
-	wire.Build(NewEvent, NewGreeter, NewMessage)
-	return Event{}, nil
+func InitializeEvent(phrase string) (event.Event, error) {
+	wire.Build(event.NewEvent, greeter.NewGreeter, message.NewMessage)
+	return event.Event{}, nil
 }
